refactor(frontend): Add ConfigService
